fix: Remove debounce and ResizeObserver
Debounce and ResizeObserver were causing an error on unmounting map where somewhere it was undefined. This error was not catched anywhere and causing error to propagate and break things.

I can have a look on it in a future again, now there is no time neither willing to spend more time on this..